move dkjson to libraries
